Move errors to vitest
